// packages/next-app/types/lenstypes.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Blockchain data scalar type */
  BlockchainData: any;
  /** Broadcast scalar id type */
  BroadcastId: any;
  /** ChainId custom scalar type */
  ChainId: any;
  /** collect module data scalar type */
  CollectModuleData: any;
  /** Contract address custom scalar type */
  ContractAddress: any;
  /** create handle custom scalar type */
  CreateHandle: any;
  /** Cursor custom scalar type */
  Cursor: any;
  /** The javascript `Date` as string. Type represents date and time as the ISO Date string. */
  DateTime: any;
  /** Ethereum address custom scalar type */
  EthereumAddress: any;
  /** follow module data scalar type */
  FollowModuleData: any;
  /** handle custom scalar type */
  Handle: any;
  /** handle claim id custom scalar type */
  HandleClaimIdScalar: any;
  /** Internal publication id custom scalar type */
  InternalPublicationId: any;
  /** jwt custom scalar type */
  Jwt: any;
  /** limit custom scalar type */
  LimitScalar: any;
  /** Markdown scalar type */
  Markdown: any;
  /** mimetype custom scalar type */
  MimeType: any;
  /** Nft ownership id type */
  NftOwnershipId: any;
  /** Nonce custom scalar type */
  Nonce: any;
  /** ProfileId custom scalar type */
  ProfileId: any;
  /** Publication id custom scalar type */
  PublicationId: any;
  /** Publication url scalar type */
  PublicationUrl: any;
  /** reference module data scalar type */
  ReferenceModuleData: any;
  /** Query search */
  Search: any;
  /** Relayer signature */
  Signature: any;
  /** Sources custom scalar type */
  Sources: any;
  /** timestamp date custom scalar type */
  TimestampScalar: any;
  /** The tx has */
  TxHash: any;
  /** UnixTimestamp custom scalar type */
  UnixTimestamp: any;
  /** Url scalar type */
  Url: any;
  /** Represents NULL values */
  Void: any;
};
// ...
export type Attribute = {
  __typename?: "Attribute";
  /** The display type */
  displayType?: Maybe<MetadataDisplayType>;
  /** identifier of this attribute, we will update by this id  */
  key: Scalars["String"];
  /** The trait type - can be anything its the name it will render so include spaces */
  traitType?: Maybe<Scalars["String"]>;
  /** Value attribute */
  value: Scalars["String"];
};
// ...
export type Dispatcher = {
  __typename?: "Dispatcher";
  /** The dispatcher address */
  address: Scalars["EthereumAddress"];
  /** If the dispatcher can use the relay */
  canUseRelay: Scalars["Boolean"];
};
// ...
export type FeeFollowModuleSettings = {
  __typename?: "FeeFollowModuleSettings";
  /** The collect module amount info */
  amount: ModuleFeeAmount;
  contractAddress: Scalars["ContractAddress"];
  /** The collect module recipient address */
  recipient: Scalars["EthereumAddress"];
  /** The follow modules enum */
  type: FollowModules;
};

export type Follow = {
  followModule?: InputMaybe<FollowModuleRedeemParams>;
  profile: Scalars["ProfileId"];
};

export type FollowModule = FeeFollowModuleSettings;
// ...
export type MediaSet = {
  __typename?: "MediaSet";
  /** Medium media - will always be null on the public API */
  medium?: Maybe<Media>;
  /** Original media */
  original: Media;
  /** Small media - will always be null on the public API */
  small?: Maybe<Media>;
};
// ...
export enum MetadataDisplayType {
  Date = "date",
  Number = "number",
  String = "string",
}
// ...
export type NftImage = {
  __typename?: "NftImage";
  /** The token image nft */
  chainId: Scalars["Int"];
  /** The contract address */
  contractAddress: Scalars["ContractAddress"];
  /** The token id of the nft */
  tokenId: Scalars["String"];
  /** The token image nft */
  uri: Scalars["Url"];
  /** If the NFT is verified */
  verified: Scalars["Boolean"];
};
// ...
export type Profile = {
  __typename?: "Profile";
  /** Optionals param to add extra attributes on the metadata */
  attributes?: Maybe<Array<Attribute>>;
  /** Bio of the profile */
  bio?: Maybe<Scalars["String"]>;
  /** The cover picture for the profile */
  coverPicture?: Maybe<ProfileMedia>;
  /** The dispatcher */
  depatcher?: Maybe<Dispatcher>;
  /** The follow module */
  followModule?: Maybe<FollowModule>;
  /** The profile handle */
  handle: Scalars["Handle"];
  /** The profile id */
  id: Scalars["ProfileId"];
  /** Is the profile default */
  isDefault: Scalars["Boolean"];
  /** Location set on profile */
  location?: Maybe<Scalars["String"]>;
  /** Metadata url */
  metadata?: Maybe<Scalars["Url"]>;
  /** Name of the profile */
  name?: Maybe<Scalars["String"]>;
  /** Who owns the profile */
  ownedBy: Scalars["EthereumAddress"];
  /** The picture for the profile */
  picture?: Maybe<ProfileMedia>;
  /** Profile stats */
  stats: ProfileStats;
  /** Twitter url set on profile */
  twitter?: Maybe<Scalars["String"]>;
  /** Website set on profile */
  website?: Maybe<Scalars["Url"]>;
};

export type ProfileMedia = MediaSet | NftImage;
// ...
export type ProfileStats = {
  __typename?: "ProfileStats";
  /** Total collects count */
  totalCollects: Scalars["Int"];
  /** Total comment count */
  totalComments: Scalars["Int"];
  /** Total follower count */
  totalFollowers: Scalars["Int"];
  /** Total following count (remember the wallet follows not profile so will be same for every profile they own) */
  totalFollowing: Scalars["Int"];
  /** Total mirror count */
  totalMirrors: Scalars["Int"];
  /** Total post count */
  totalPosts: Scalars["Int"];
  /** Total publication count */
  totalPublications: Scalars["Int"];
};